First pass at using an Application class to build a Phosphide app with its chrome.
