+++ LICENSE
@@ -1,4 +1,4 @@
-Copyright (c) 2014-2015, PhosphorJS Contributors
+Copyright (c) 2014-2016, PhosphorJS Contributors
 All rights reserved.
 
 Redistribution and use in source and binary forms, with or without
