+++ example/application/tsconfig.json
@@ -1,0 +1,13 @@
+{
+  "compilerOptions": {
+    "noImplicitAny": true,
+    "noEmitOnError": true,
+    "module": "commonjs",
+    "moduleResolution": "node",
+    "target": "ES5"
+  },
+  "files": [
+    "../../typings/es6-promise/es6-promise.d.ts",
+    "index.ts"
+  ]
+}
